Spherical almost 100% done
Now, we are left to fix the bugs
- the first one is that the lines on the background are showing (they
  should not)

+++ includes/methods_and_classes_bonus/t_map_bonus.h
@@ -6,7 +6,7 @@
 /*   By: vcedraz- <vcedraz-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/28 09:14:16 by vcedraz-          #+#    #+#             */
-/*   Updated: 2023/02/04 19:01:03 by vcedraz-         ###   ########.fr       */
+/*   Updated: 2023/02/05 15:35:50 by vcedraz-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,20 +31,20 @@ typedef struct s_mp
 	int					is_vertical;
 	int					is_square;
 	int					is_plateau;
-	int					width;
-	int					height;
+	double				width;
+	double				height;
 	uint				size;
 	short				max_z;
 	short				min_z;
-	double	 			max_x;
-	double	 			min_x;
-	double	 			max_y;
-	double	 			min_y;
+	double				max_x;
+	double				min_x;
+	double				max_y;
+	double				min_y;
 	double				initial_max_z;
 	double				target_width;
 	double				target_height;
 	double				ratio;
-	double 				radius;
+	double				radius;
 	t_n					*pts;
 	t_point				**arr;
 	t_point				**(*create_arrmap)(t_map *map);
